### 2020-03-01 #2131 @labarba
👋  @poulson — This submission falls in the general field of computer vision, where we don't have any topic editors. It runs a global nonlinear optimization to minimize the error between the projected location of 3D points in each camera and the actual observed location of the point in the camera frames. Could you handle it as editor?
