### 2020-03-09 #2131 @poulson
Hi @fiveop :wave: -- given your preferred language of Rust, would you feel comfortable serving as one of the reviewers of this paper on a Rust software package for generating datasets for bundle adjustment problems?
