### 2020-03-09 #2131 @poulson
Hi @fhorrobin :wave: -- given that you mentioned Rust as one of your preferred languages, and your stated interest in statistical modeling, would you be interested and able to serve as one of the reviewers for this submission on a Rust package for generating datasets for bundle adjustment problems?
