### 2020-03-23 b89b137
Format & close blocks to form full grid
