+++ src/generate.rs
@@ -388,7 +388,7 @@ where
 
     let mut points = Vec::with_capacity(num_points);
     let mut fail_count = 0;
-    let fail_threshold = 10*num_points;
+    let fail_threshold = 10 * num_points;
     while points.len() < num_points && fail_count < fail_threshold {
         let i = dist.sample(&mut rng);
         let (m, j) = indices[i];
@@ -408,7 +408,12 @@ where
     }
 
     if fail_count >= fail_threshold {
-        panic!("Failed to generate enough points. {} successes, {} failures, {} requested points.", points.len(), fail_count, num_points);
+        panic!(
+            "Failed to generate enough points. {} successes, {} failures, {} requested points.",
+            points.len(),
+            fail_count,
+            num_points
+        );
     }
 
     points
