### 2020-04-06 fd04e7b
Better error messages on invalid input
