### 2020-06-01 #2131 @poulson
Hi Fergus (@fhorrobin) :wave: -- given that you have listed Rust as one of your preferred programming language, as well as expertise in statistical methods and solid body dynamics, you appear to be one of the most qualified reviewers in our list. I'm fully aware that now is a very unstable time, and you could likely be overwhelmed, but please let me know if you would have the time to review this submission, as it has been on hold for a few months.

Hi Sameer (@sandwichmaker) :wave: -- given the PDF and, if we get a Rust specialist as a reviewer, would you feel comfortable reviewing the mathematics and methodology of this submission?
