### 2020-07-24 #2131 @poulson
Hi Conrad (@conradsnicta) :wave: -- You've expressed interest in reviewing JOSS submissions and have listed both computer vision as one of your specialties. While the submission is in Rust, would you feel comfortable helping to review this submission?
