### 2020-07-24 #2131 @poulson
Hi Valentin (@vchuravy) :wave: -- Like Conrad, you have maked computer vision as one of your specialties and listed a primary language that is different from this submision's (Julia vs. Rust). Would you feel comfortable helping to review this submission on bundle adjustment in Rust?
