### 2020-07-24 #2131 @poulson
Hi Abhishek (@abhishekbajpayee) :wave: -- You have mentioned computer vision as one of your specialties and have listed both Python and C++ as preferred programming languages. Would you feel comfortable reviewing this submission on bundle adjustment problems in Rust?
