### 2020-08-12 #2131 @conradsnicta
@poulson Apologies for the late response. Been very busy.  Based on the contents of the paper draft, I currently don't have sufficient knowledge in the applicable areas (SfM and SLAM).
