### 2020-09-05 #2131 @poulson
Hi Mathieu :wave:  (@matbesancon) -- I hate to repeatedly request you as a reviewer, but, as you can see, we've been having issues finding reviewers who are comfortable with both Rust and reviewing a bundle adjustment project. You list both Rust and a combination of optimization and statistics in your expertise and are one of the better first. Would you happen to be interested and able to help review this submission?
