### 2020-09-05 #2131 @poulson
Hi Lachlan :wave: (@breezykermo) -- Given your listed expertise in Rust and computer graphics, you appear to be one of the few potential reviewers who fits the bill for this Rust Bundle Adjustment submission. Would you be interested and able to help review this submission?
