### 2020-09-07 #2131 @breezykermo
Hi @poulson; although I do not have any expertise in BA, I would be happy to help review if you think it would be useful. What exactly would a review require from me? I have not reviewed or contributed through JOSS yet. :)
