### 2020-10-05 #2131 @kthyng
Hi @breezykermo! I am the rotating associate editor in chief this week so I can help with your question. JOSS reviews are done as a github issue (which will be opened separately from this pre-review issue). They are checklist-driven, and you can get more information on the process here: https://joss.readthedocs.io/en/latest/reviewer_guidelines.html. We basically want reviewers to use the software as a new user from installation, going through docs, and verifying it works as it says it should. Issues and comments that come up should be documented as issues on the software repository and linked to the review issue by mentioning it inline.

Let me know if you have any other questions!
