### 2020-10-05 #2131 @kthyng
> @poulson I'm new to the joss community, so I don't know anyone who make a good reviewer.

@tkonolige as you can see, @poulson is having a hard time finding reviewers for your submission. I wouldn't say that any familiarity with the JOSS community is needed to suggest reviewers, just like it isn't for a traditional academic paper. We provide a list of possible reviewers you can [peruse](https://bit.ly/joss-reviewers) to see if anyone seems like a fit (if so, please list them here without using "@" so they aren't directly pinged). Or, perhaps you know people in your scientific community and network and can provide contact information for them, or ask them for potential reviewers.
