### 2020-12-03 #2131 @adi3
@poulson I work in robotics, and have experience with SLAM and bundle adjustments. I'd be happy to take this on. Let me know! I do not have hands-on experience in Rust, but can review the code against common best practices for software and documentation.
