### 2020-12-07 #2131 @adi3
@breezykermo since JOSS papers require 2 reviewers, I would suggest we both do it. I have reviewed for JOSS in the past and can guide you through the review process, although I think the documentation is already quite helpful with this.
@kthyng @poulson your call :)
