### 2021-03-22 #3125 @whedon
[REVIEW]: City2BA: Tools for creating synthetic bundle adjustment problems
**Submitting author:** @tkonolige (<a href="http://orcid.org/0000-0002-5052-6479">Tristan Konolige</a>)
**Repository:** <a href="https://github.com/tkonolige/city2ba" target ="_blank">https://github.com/tkonolige/city2ba</a>
**Version:** v1.1.0
