### 2021-04-12 #3125 @adi3
@tkonolige Please fix invalid DOIs and then run `@whendon check references` to make sure your DOIs list as OK.
