+++ paper.bib
@@ -20,7 +20,7 @@
   booktitle={Computer Vision – ECCV 2014},
   pages={61--75},
   year={2014},
-  doi={https://doi.org/10.1007/978-3-319-10578-9_5},
+  doi={10.1007/978-3-319-10578-9_5},
   organization={Springer}
 }
 @inproceedings{klingner2013street,
@@ -28,7 +28,7 @@
   author={Klingner, Bryan and Martin, David and Roseborough, James},
   booktitle={Proceedings of the IEEE International Conference on Computer Vision},
   pages={953--960},
-  doi={https://doi.org/10.1109/iccv.2013.122},
+  doi={10.1109/iccv.2013.122},
   year={2013}
 }
 
