### 2021-09-10 #3125 @tkonolige
@adi3 I have some example usage in the readme. Is there something missing from it that you would like you would like to see? Right now you have to provide your own .obj file, but I could have the example download one.
