### 2022-02-21 #3125 @adi3
@tkonolige could you please add an example `.obj` file in the repository so users can test the tool right out-of-the-box? Thanks!
