### 2022-03-15 #3125 @adi3
@tkonolige having trouble installing the tool. Cargo is existing with this error:

```
note: /usr/bin/ld: cannot find -lembree3
          collect2: error: ld returned 1 exit status

error: failed to compile `city2ba v1.1.0`, intermediate artifacts can be found at `/tmp/cargo-installKuVNju`

Caused by:
  could not compile `city2ba` due to previous error
```

I'm on Ubuntu LTS 20.04. Can you please see what might be wrong? Either you can fix this, or perhaps you could let me know which exact platform you tested your software on?
