### 2022-04-25 #3125 @adi3
@tkonolige - here's my feedback so far:

- I installed `embree` on a fresh Linux box and adding the env var like you suggested indeed did the trick. I would recommend you add both these things to the installation instructions in a code block. Or at least the env export part.
- Is there a way to simply this and install `embree` directly via cargo?
- After installation, cargo asks the user to add the bin to their env `PATH`. Could you please add that step in the installation guide please? I believe it would be a simple `export PATH=$PATH:/home/ubuntu/.cargo/bin`.
- In the functionality testing, please add a quick line about downloading the `test_scene.obj`.
- It seems the program API has changed (`--num-cameras` to `--cameras`, `--num-points` to `--points`, etc). Can you please update the documentation to reflect these changes?
- I get this error when trying to run `generate` on the test_scene object: `Cannot generate world points with 0 cameras.` Can you please look into it and tell me the fixes? Here's the stacktrace:
```
> city2ba generate test_scene.obj problem.bal --cameras 100 --points 200
Generated 0 cameras
Modified intrinsics
thread 'main' panicked at 'Cannot generate world points with 0 cameras.', /home/ubuntu/.cargo/registry/src/github.com-1ecc6299db9ec823/city2ba-1.1.0/src/generate.rs:366:9
stack backtrace:
   0: std::panicking::begin_panic
   1: city2ba::generate::generate_world_points_uniform
   2: city2ba::main
note: Some details are omitted, run with `RUST_BACKTRACE=full` for a verbose backtrace.
```
