### 2022-06-08 #3125 @tkonolige
@adi3
> * I installed `embree` on a fresh Linux box and adding the env var like you suggested indeed did the trick. I would recommend you add both these things to the installation instructions in a code block. Or at least the env export part.
Added
> * Is there a way to simply this and install `embree` directly via cargo?
Yes, but it is not easy and would require work in an upstream dependency.
> * After installation, cargo asks the user to add the bin to their env `PATH`. Could you please add that step in the installation guide please? I believe it would be a simple `export PATH=$PATH:/home/ubuntu/.cargo/bin`.
Done
> * In the functionality testing, please add a quick line about downloading the `test_scene.obj`.
Done
> * It seems the program API has changed (`--num-cameras` to `--cameras`, `--num-points` to `--points`, etc). Can you please update the documentation to reflect these changes?
Done
> * I get this error when trying to run `generate` on the test_scene object: `Cannot generate world points with 0 cameras.` Can you please look into it and tell me the fixes? Here's the stacktrace:
Looks like not enough locations for cameras were generated, try increasing the number for `--cameras`. (I've added a statement to that effect to the error message).
