### 2022-09-18 #3125 @adi3
@tkonolige - I'm almost finished reviewing this submission. Functionality-wise, I'm satisfied. However, in your paper, we need a section titled "Statement of Need" which will include most of the information you already have in your first paragraph. The Summary section itself can just be a few lines long and should provide only a high-level overview of the software and its benefits. Once you make these quick changes, I'm happy to check off the last few items.
