### 2022-11-27 #3125 @arfon
👋 @tkonolige – just checking in again here. Can you confirm you're still interested in publishing this paper in JOSS?
