### 2023-01-10 #3125 @arfon
@tkonolige – I made some suggested edits to your paper – could you take a look at let me know what you think? https://github.com/tkonolige/city2ba/pull/5
